22.11.2023
Sum up the distance of each angle inside a sliding window, then choose the best window as the gap to follow. The steering angle is calculated based on: 1. The furthest gap, 2. The closest distance to the obstacles. 
For improvement: 1. Add the distance to the middle line, 2. Adjust window size according to the distance/speed/shape of lidar vector input. 3. Adjust speed according to distance/angle